Added Sasl as a part of Standard Connection
