+++ csharp/src/Drivers/Apache/Hive2/HiveServer2TransportType.cs
@@ -21,14 +21,12 @@ namespace Apache.Arrow.Adbc.Drivers.Apache.Hive2
     {
         Http,
         Standard,
-        Sasl,
         Empty = int.MaxValue,
     }
 
     internal static class HiveServer2TransportTypeParser
     {
-        internal const string SupportedList = HiveServer2TransportTypeConstants.Http + ", " +
-            HiveServer2TransportTypeConstants.Standard + ", " + HiveServer2TransportTypeConstants.Sasl;
+        internal const string SupportedList = HiveServer2TransportTypeConstants.Http + ", " + HiveServer2TransportTypeConstants.Standard;
 
         internal static bool TryParse(string? serverType, out HiveServer2TransportType serverTypeValue)
         {
@@ -43,9 +41,6 @@ namespace Apache.Arrow.Adbc.Drivers.Apache.Hive2
                     return true;
                 case HiveServer2TransportTypeConstants.Standard:
                     serverTypeValue = HiveServer2TransportType.Standard;
-                    return true;
-                case HiveServer2TransportTypeConstants.Sasl:
-                    serverTypeValue = HiveServer2TransportType.Sasl;
                     return true;
                 default:
                     serverTypeValue = default;
